Add :Rschema command
When calling :Rschema with a given table name as an argument, Vim will
jump to the table's position in the db/schema.rb file. When no argument
is given :Rschema will use the current file's model's table name, and
jump to it's position in db/schema.rb. Similar to how :R behaves when
called inside a model file.